Ported vancouver to x86_64 and improved the cross-compiler Now we use gcc 4.6.3 and we support -mcmodel=large, which is important for x86_64, because it allows us to put code and data above 2G.

+++ nre/apps/vancouver/Vancouver.cc
@@ -50,7 +50,7 @@ PARAM_HANDLER(m, "m - specify the amount of memory for the guest in MiB") {
 			DataSpaceDesc::RWX | DataSpaceDesc::BIGPAGES);
 }
 PARAM_HANDLER(vcpus, " vcpus - instantiate the vcpus defined with 'ncpu'") {
-	for(unsigned count = 0; count < ncpu; count++)
+	for(size_t count = 0; count < ncpu; count++)
 		mb.parse_args("vcpu halifax vbios lapic");
 }
