+++ nre/apps/vancouver/bus/bus.h
@@ -59,8 +59,8 @@ class DBus {
 	};
 
 	unsigned long _debug_counter;
-	unsigned _list_count;
-	unsigned _list_size;
+	size_t _list_count;
+	size_t _list_size;
 	struct Entry *_list;
 
 	/**
@@ -69,7 +69,7 @@ class DBus {
 	DBus(const DBus<M> &bus);
 	DBus& operator=(const DBus<M> &bus);
 
-	void set_size(unsigned new_size) {
+	void set_size(size_t new_size) {
 		Entry *n = new Entry[new_size];
 		if(_list) {
 			memcpy(n,_list,_list_count * sizeof(*_list));
@@ -94,7 +94,7 @@ public:
 	bool send(M &msg,bool earlyout = false) {
 		_debug_counter++;
 		bool res = false;
-		for(unsigned i = _list_count; i-- && !(earlyout && res);)
+		for(size_t i = _list_count; i-- && !(earlyout && res);)
 			res |= _list[i]._func(_list[i]._dev,msg);
 		return res;
 	}
@@ -105,7 +105,7 @@ public:
 	bool send_fifo(M &msg) {
 		_debug_counter++;
 		bool res = false;
-		for(unsigned i = 0; i < _list_count; i++)
+		for(size_t i = 0; i < _list_count; i++)
 			res |= _list[i]._func(_list[i]._dev,msg);
 		return 0;
 	}
@@ -116,7 +116,7 @@ public:
 	 */
 	bool send_rr(M &msg,unsigned &start) {
 		_debug_counter++;
-		for(unsigned i = 0; i < _list_count; i++) {
+		for(size_t i = 0; i < _list_count; i++) {
 			if(_list[i]._func(_list[(i + start) % _list_count]._dev,msg)) {
 				start = (i + start + 1) % _list_count;
 				return true;
@@ -128,7 +128,7 @@ public:
 	/**
 	 * Return the number of entries in the list.
 	 */
-	unsigned count() {
+	size_t count() {
 		return _list_count;
 	}
 
@@ -137,7 +137,7 @@ public:
 	 */
 	void debug_dump() {
 		nre::Serial::get().writef("%s: Bus used %ld times.",__PRETTY_FUNCTION__,_debug_counter);
-		for(unsigned i = 0; i < _list_count; i++) {
+		for(size_t i = 0; i < _list_count; i++) {
 			nre::Serial::get().writef("\n%2d:\t",i);
 			_list[i]._dev->debug_dump();
 		}
