+++ nre/apps/vancouver/model/memorycontroller.cc
@@ -25,8 +25,8 @@ using namespace nre;
 
 class MemoryController : public StaticReceiver<MemoryController> {
 	char *_physmem;
-	unsigned long _start;
-	unsigned long _end;
+	uintptr_t _start;
+	uintptr_t _end;
 
 public:
 	/****************************************************/
@@ -53,7 +53,7 @@ public:
 		return true;
 	}
 
-	MemoryController(char *physmem,unsigned long start,unsigned long end)
+	MemoryController(char *physmem,uintptr_t start,uintptr_t end)
 		: _physmem(physmem), _start(start), _end(end) {
 	}
 };
@@ -65,8 +65,8 @@ PARAM_HANDLER(mem,
 	MessageHostOp msg(MessageHostOp::OP_GUEST_MEM,0UL);
 	if(!mb.bus_hostop.send(msg))
 		Util::panic("%s failed to get physical memory\n",__PRETTY_FUNCTION__);
-	unsigned long start = ~argv[0] ? argv[0] : 0;
-	unsigned long end = argv[1] > msg.len ? msg.len : argv[1];
+	uintptr_t start = ~argv[0] ? argv[0] : 0;
+	uintptr_t end = argv[1] > msg.len ? msg.len : argv[1];
 	Serial::get().writef("physmem: %lx %p [%lx, %lx]\n",msg.value,msg.ptr,start,end);
 	MemoryController *dev = new MemoryController(msg.ptr,start,end);
 	// physmem access
